Feature #16: separate component
